fixup: restore integral parameter types
The automatic code translation tool erroneously normalizes integer types
of sizes less that int to int. This implicated char, short and
typedefs of such.

+++ dsap/common/attrv.c
@@ -105,7 +105,7 @@ char	multiline;
 }
 
 int
-set_av_pe_print (int sntx, char *print_pe) {
+set_av_pe_print (short sntx, char *print_pe) {
 	if (sntx >= num_syntax)
 		return;
 
@@ -165,7 +165,7 @@ short x;
 }
 
 char *
-syntax2str (int sntx) {
+syntax2str (short sntx) {
 	if (sntx >= num_syntax)
 		return NULLCP;
